Tests/LibWeb/Text/expected/DOM/shadow-root-boundary-of-inserted-node-is-traversed.txt

Hello from script in the shadow root of the just inserted div!
